Separating GSMConnectionManager definition/implementation in a h/cpp file
